feat: New, experimental tokenizer
The previous tokenizer/parser was implemented in an ad-hoc way. With the
knowledge gained from doing it the first time, let's do it again but
with the POSIX spec in mind from the start.

This commit introduces the new lexer, which is quite similar to the
previous one. Though it's based more on POSIX, resulting in support for
builtin keywords and similar.